+++ packages/astro/src/content/loaders/glob.ts
@@ -1,10 +1,10 @@
 import { fileURLToPath, pathToFileURL } from 'url';
-import type { Loader, LoaderContext } from './sync.js';
+import type { Loader, LoaderContext } from './types.js';
 import { promises as fs } from 'fs';
 import fastGlob from 'fast-glob';
 import pLimit from 'p-limit';
-import { getContentEntryIdAndSlug, getEntryConfigByExtMap } from './utils.js';
-import type { ContentEntryType } from '../@types/astro.js';
+import { getContentEntryIdAndSlug, getEntryConfigByExtMap } from '../utils.js';
+import type { ContentEntryType } from '../../@types/astro.js';
 import micromatch from 'micromatch';
 import { relative } from 'path/posix';
 import { green } from 'kleur/colors';
