+++ packages/astro/src/content/loaders/index.ts
@@ -1,0 +1,3 @@
+export { file } from './file.js';
+export { glob } from './glob.js';
+export * from './types.js';
