+++ packages/astro/test/fixtures/content-layer/src/loaders/post-loader.ts
@@ -1,4 +1,5 @@
-import { type Loader, z } from 'astro:content';
+import { z } from 'astro:content';
+import type { Loader } from "astro/loaders"
 
 export interface PostLoaderConfig {
 	url: string;
